+++ src/main/java/baseball/Application.java
@@ -1,6 +1,6 @@
 package baseball;
 
-import static baseball.BaseballGame.playBaseballGame;
+import static baseball.game.BaseballGame.playBaseballGame;
 
 public class Application {
     public static void main(String[] args) {
