+++ src/main/java/baseball/exception/RestartNumberValidator.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.exception;
 
 public class RestartNumberValidator {
