+++ src/main/java/baseball/exception/UserNumberValidator.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.exception;
 
 import java.util.List;
